[Single Object state maintenance] Use updates that comprises of both name and message
The point to be noted here is that the presenter component need not
be changed even though we modify the way data is passed around i.e.,
instead of passing around name and message we pass updates object
which comprises of name and message but still props.name and props.message
remain as it is. The reason for this is because the props of the
presentation component (Greetermessage) are the props of the Container
component(Greeter) which are updated as part of 'updates'.

Signed-off-by: srikanth <srikanth.chevalam@visteon.com>